+++ cpp/l2-proxy/l2-proxy-version.h
@@ -1,4 +1,4 @@
 #ifndef L2_PROXY_VERSION_H
 #define L2_PROXY_VERSION_H
-#define VERSION "1.0.4-bd49d90"
+#define VERSION "1.0.4-377f7ba"
 #endif